Make hill definition configurable in the namelist
This version defines the N-S risge, and makes the height and width of the ridge configurable in the namelist.

We add hill_height and hill_width to the &domains block in the namelist. Both are defined in units of metres.

Do a ./clean -a; configure; compile em_les to implement.